requirements.txt
Django>=5.0,<5.1
Pillow>=10.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
reportlab>=4.0.0
qrcode>=7.4.2
